Follow-up to r17402 - Fix compiling warnings and VCC errors
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@17403 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/status.c
@@ -1906,7 +1906,7 @@ static unsigned short status_base_atk(const struct block_list *bl, const struct 
 #ifdef RENEWAL
 unsigned int status_weapon_atk(struct weapon_atk wa, struct status_data *status)
 {
-	float str = status->str;
+	short str = status->str;
 
 	if (wa.range > 1)
 		str = status->dex;
